Move all the fast stuff into src/fast
