+++ setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup, find_packages
+
+setup(
+    name='fast',
+    version='0',
+    package_dir={'': 'src'},
+    packages=find_packages(where='src')
+)
