+++ src/fast/analysis/base_analysis.py
@@ -11,7 +11,7 @@
 
 
 import numpy as np
-from ..base import base
+from fast.base import base
 
 
 #######################################################################
