+++ src/fast/analysis/contacts.py
@@ -11,13 +11,13 @@
 #######################################################################
 
 
-import gc
 import glob
+import itertools
 import mdtraj as md
 import numpy as np
 import os
-from .base_analysis import base_analysis
-from .. import tools
+from fast.analysis.base_analysis import base_analysis
+from fast import tools
 
 
 #######################################################################
@@ -25,8 +25,61 @@ from .. import tools
 #######################################################################
 
 
-class RMSDWrap(base_analysis):
-    """Analysis wrapper for calculating state rmsds
+def best_hummer_q(traj, native, verbose=False, native_cutoff=0.45):
+    """Compute the fraction of native contacts according the definition from
+    Best, Hummer and Eaton [1].
+    Adapted from: 'http://mdtraj.org/latest/examples/native-contact.html'
+    
+    Parameters
+    ----------
+    traj : md.Trajectory
+        The trajectory to do the computation for
+    native : md.Trajectory
+        The 'native state'. This can be an entire trajecory, or just a single frame.
+        Only the first conformation is used
+        
+    Returns
+    -------
+    q : np.array, shape=(len(traj),)
+        The fraction of native contacts in each frame of `traj`
+        
+    References
+    ----------
+    ..[1] Best, Hummer, and Eaton, "Native contacts determine protein folding
+          mechanisms in atomistic simulations" PNAS (2013)
+    """
+
+    BETA_CONST = 50  # 1/nm
+    LAMBDA_CONST = 1.8
+    NATIVE_CUTOFF = native_cutoff  # nanometers
+
+    # get the indices of all of the heavy atoms
+    heavy = native.topology.select_atom_indices('heavy')
+    # get the pairs of heavy atoms which are farther than 3
+    # residues apart
+    heavy_pairs = np.array(
+        [(i,j) for (i,j) in itertools.combinations(heavy, 2)
+            if abs(native.topology.atom(i).residue.index - \
+                   native.topology.atom(j).residue.index) > 3])
+
+    # compute the distances between these pairs in the native state
+    heavy_pairs_distances = md.compute_distances(native[0], heavy_pairs)[0]
+    # and get the pairs s.t. the distance is less than NATIVE_CUTOFF
+    native_contacts = heavy_pairs[heavy_pairs_distances < NATIVE_CUTOFF]
+    if verbose:
+        print("Number of native contacts", len(native_contacts))
+
+    # now compute these distances for the whole trajectory
+    r = md.compute_distances(traj, native_contacts)
+    # and recompute them for just the native state
+    r0 = md.compute_distances(native[0], native_contacts)
+    q = np.mean(
+        1.0 / (1 + np.exp(BETA_CONST * (r - LAMBDA_CONST * r0))), axis=1)
+    return q
+
+
+class ContactsWrap(base_analysis):
+    """Analyses the fraction of native contacts.
 
     Parameters
     ----------
@@ -60,7 +113,7 @@ class RMSDWrap(base_analysis):
 
     @property
     def class_name(self):
-        return "RMSDWrap"
+        return "ContactsWrap"
 
     @property
     def config(self):
@@ -75,7 +128,7 @@ class RMSDWrap(base_analysis):
 
     @property
     def base_output_name(self):
-        return "rmsd_per_state"
+        return "contacts_per_state"
 
     def run(self):
         # determine if file already exists
@@ -91,7 +144,7 @@ class RMSDWrap(base_analysis):
                 struct_sub = self.base_struct_md
             else:
                 struct_sub = self.base_struct_md.atom_slice(self.atom_indices_vals)
-            # calculate and save rmsds
-            rmsds = md.rmsd(centers, struct_sub)
-            np.save(self.output_name, rmsds)
+            # calculate and save contacts
+            contacts = best_hummer_q(centers, struct_sub)
+            np.save(self.output_name, contacts)
