+++ src/fast/analysis/distances.py
@@ -18,8 +18,8 @@ import mdtraj as md
 import numpy as np
 import os
 import sys
-from .base_analysis import base_analysis
-from .. import tools
+from fast.analysis.base_analysis import base_analysis
+from fast import tools
 
 
 #######################################################################
