+++ src/fast/exception.py
