+++ src/fast/md_gen/upside.py
@@ -13,9 +13,9 @@ import mdtraj as md
 import numpy as np
 import os
 import mdtraj_upside as mu
-from .. import tools
-from .. import submissions
-from ..base import base
+from fast import tools
+from fast import submissions
+from fast.base import base
 
 #######################################################################
 # code
