+++ src/fast/tools.py
